Worked on dif_ekf_p1().
Turns out measurementFcn() needs to handle complex numbers; refactored it to do so.
Added zbound() helper to round really small values to zero. Matlab seems to do this automatically. We use zbound() to fix a problem in the measurementFcn() where there is a sqrt() calculation on a really small complex number. Normally, we can leave small numbers as is, but because of the sqrt(), the sign of the result is affected and that could have consequences on future multiplications.
Also discovered that the Matlab code works with seconds as the unit for timestamps.